pyproject.toml
[tool.poetry]
name = "falcon"
version = "0.1.0"
description = ""
authors = ["xolmomin <xolmomin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-crispy-forms = "^2.0"
pillow = "^9.5.0"
django-import-export = "^3.1.0"
django-mptt = "^0.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
